create empresa

+++ db/database.sql
@@ -62,7 +62,7 @@ CREATE TABLE `gestiondb`.`direccion` (
 CREATE TABLE `gestiondb`.`empresa` (
   `idempresa` INT NOT NULL AUTO_INCREMENT,
   `nombre` VARCHAR(45) NOT NULL,
-  `telefono` INT NOT NULL,
+  `telefono` BIGINT NOT NULL,
   `cuit` BIGINT NOT NULL,
   `iddireccion` INT NOT NULL,
   PRIMARY KEY (`idempresa`),
